Added indexes for performance efficiencies

+++ db/migrate/20210819110747_create_audit_trails.rb
@@ -1,10 +1,10 @@
 class CreateAuditTrails < ActiveRecord::Migration[6.1]
   def change
     create_table :audit_trails do |t|
-      t.integer :record_id
+      t.bigint :record_id
       t.string :record_type, limit: 30
       t.string :event, limit: 20
-      t.integer :user_id
+      t.bigint :user_id
       t.datetime :created_at
     end
   end
